+++ app/components/admin/forms/base_component.rb
@@ -1,0 +1,13 @@
+module Admin
+  module Forms
+    class BaseComponent < ApplicationComponent
+      attr_reader :form, :field, :classes
+
+      def initialize(form, field, classes:)
+        @form = form
+        @field = field
+        @classes = classes
+      end
+    end
+  end
+end