+++ app/components/admin/tables/cell_alignment.rb
@@ -1,0 +1,9 @@
+module Admin
+  module Tables
+    module CellAlignment
+      def align_class
+        align == :right ? "text-right" : "text-left"
+      end
+    end
+  end
+end