+++ app/components/admin/tables/row_component.rb
@@ -1,0 +1,7 @@
+module Admin
+  module Tables
+    class RowComponent < ApplicationComponent
+      renders_many :cells, CellComponent
+    end
+  end
+end